+++ csatar-plugins/csatar/models/GalleryModelPivot.php
@@ -33,4 +33,8 @@ class GalleryModelPivot extends Model
         'model_type',
     ];
 
+    public $morphTo = [
+        'model' => []
+    ];
+
 }