+++ csatar-plugins/csatar/models/OrganizationBase.php
@@ -219,4 +219,8 @@ class OrganizationBase extends PermissionBasedAccess
         }
         return $this->original[$attribute];
     }
+
+    public function getParentTree() {
+        return null;
+    }
 }